Fixed review comments
1. Handled change in apply_additional_filters signature
2. Simplified ProjectStat model
3. Removed unwanted guard clause from Temp Index
4. Added dev_and_test_envs check along with .com? check

+++ db/post_migrate/20220622080547_backfill_project_statistics_with_container_registry_size.rb
@@ -12,7 +12,7 @@ class BackfillProjectStatisticsWithContainerRegistrySize < Gitlab::Database::Mig
   disable_ddl_transaction!
 
   def up
-    return unless Gitlab.com?
+    return unless Gitlab.dev_or_test_env? || Gitlab.com?
 
     queue_batched_background_migration(
       MIGRATION_CLASS,
@@ -26,7 +26,7 @@ class BackfillProjectStatisticsWithContainerRegistrySize < Gitlab::Database::Mig
   end
 
   def down
-    return unless Gitlab.com?
+    return unless Gitlab.dev_or_test_env? || Gitlab.com?
 
     delete_batched_background_migration(MIGRATION_CLASS, :container_repositories, :project_id, [])
   end
